+++ osidb/serializer.py
@@ -1268,6 +1268,7 @@ class AffectSerializer(
                 "delegated_resolution",
                 "cvss_scores",
                 "purl",
+                "subpackage_purls",
                 "not_affected_justification",
                 "delegated_not_affected_justification",
                 "resolved_dt",