+++ osidb/__init__.py
@@ -2,5 +2,5 @@
 osidb version
 """
 
-__version__ = "5.2.2"
+__version__ = "5.3.0"
 VERSION = __version__